+++ src/components/auth/onboarding/index.ts
@@ -1,5 +1,4 @@
 export { ProfileCompletionStep } from "./ProfileCompletionStep";
 export { SubscriptionSelectionStep } from "./SubscriptionSelectionStep";
 export { CreateOrganizationStep } from "./CreateOrganizationStep";
-export { LicenseModal } from "./LicenseModal";
 export { PasskeySetupStep } from "./PasskeySetupStep";