+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [2.3.0-rc.3](https://github.com/janovix/auth/compare/v2.3.0-rc.2...v2.3.0-rc.3) (2026-02-28)
+
+
+### Features
+
+* add Radix UI components and new Tax & Compliance features ([6a9979c](https://github.com/janovix/auth/commit/6a9979ca7709f59ab7de4ab48ecfb761aa1bfd22))
+
+# [2.3.0-rc.2](https://github.com/janovix/auth/compare/v2.3.0-rc.1...v2.3.0-rc.2) (2026-02-28)
+
+
+### Features
+
+* **billing:** update watchlist query limits and add subscription status handling ([1be8f7d](https://github.com/janovix/auth/commit/1be8f7d0b572b0a4f7d3c78169cb2f18f4b999ec))
+
+# [2.3.0-rc.1](https://github.com/janovix/auth/compare/v2.2.1-rc.1...v2.3.0-rc.1) (2026-02-26)
+
+
+### Features
+
+* **notifications:** add notification sound preferences to settings ([339534d](https://github.com/janovix/auth/commit/339534d2b384cf34741096f6c19dd42b00c3d492))
+
 ## [2.2.1](https://github.com/janovix/auth/compare/v2.2.0...v2.2.1) (2026-02-26)
 
 
@@ -5,6 +26,7 @@
 
 * **auth:** update autoComplete attribute in LoginView for email input ([33608e8](https://github.com/janovix/auth/commit/33608e8e2cdf59c8c9811c34eec7e4b81c9f8443))
 * **layout:** update settings link in AppSwitcher component ([b03303a](https://github.com/janovix/auth/commit/b03303a47dc9ab09ba2c9b8c7ab8f8447dd00a90))
+
 
 ## [2.2.1-rc.1](https://github.com/janovix/auth/compare/v2.2.0...v2.2.1-rc.1) (2026-02-26)